+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - updated readme
 - for OS X, also check if 'reattach-to-user-namespace' is installed
+- add a feature for copying current command line to clipboard
 
 ### v0.0.1, Jun 24, 2014